Delete duplicate configuration files
For some projects both App.config and app.config were present, with
content that differed only by whitespace.
This would work on case-sensitive file system but not on a
case-insensitive file system (such as is used in Windows).
The project files referenced App.config (which would then be the file
loaded into the project and built from for case-sensitive file systems)
but on case-insensitive file systems, app.config was retained (git
overwrites the App.config with the app.config). So for a
case-insensitive file system, a different copy of the configuration file
would be used (as App.config in the project file matches app.config on
disk).
This would work until someone forgot to update both files when making
changes, then the results would differ between case-sensitive and
case-insensitive file systems.