Add .NET 8 as a target
Upgrades the projects to allow building for net8.0-windows and
netstandard2.0.
This is intended to allow building for either .NET or .NET Framework.
As the CefSharp project distributes different libraries for .NET
Framework and .NET (for example CefSharp.Common and CefSharp.NetCore)
the project files need to include different package references
depending on which framework is being target.
The nuget package will then include the correct package reference for
each target framework.