Update project files to use package references
During rework to include .NET 8 as a target framework, the project files
were edited to use project references to improve development workflow.
Now that the version is approaching ready for release, the project files
need to be updated to point to the nuget packages, not the code in the
working tree.

Also updated packages to have README in markdown format for display on
nuget.org.