Make release notes in package a link
Rather than maintaining the release notes on the github releases page
and as text in the project file XML, make the nuget package release note
tag a link to the github releases page.
Also added links to 10Duke Enterprise documentation and the release
notes to the repository/package README files.