Update release notes tag in project file
Makes the release notes tag for the nuget package release notes a URL to
the releases page on github rather than text embedded in the project
file.
This should mean that the tag does not need to be updated for each
release, the person reading the nuget package page on nuget.org can
simply go to the release notes for the specific release on the github
releases page.