Update Chromium Embedded Framework (CEF) and CefSharp dependencies
Chromium Embedded Framework (CEF) is the native binary of the C++ source code
for the Chromium engine, while CefSharp provides the .NET bindings that expose CEF
to .NET applications. These dependencies must always be updated together, especially
when the major version changes, as it is tied to the Chrome/Chromium major version.

Transient dependencies brought from the parent libraries of WPF and WinForms
where removed from the samples since they are already included.